Mudblood.Core: Unify error handling
Make MB an instance of MonadError StackTrace, where StackTrace is
a list ("stack") of (String, String) pairs. The first string denotes the
subsystem that generated the error, the second is the error messages.
Functions in the MB Monad can throw, catch and forward errors, where
forwarding means that another messages is pushed onto the stack.